refactor: moves transformMarkdownFile into markdown.ts actions
