+++ src/lib/actions/notes.ts
@@ -2,15 +2,6 @@
 import { Note, NoteFrontmatter, Topic } from "@/types";
 import fs from "fs";
 import matter from "gray-matter";
-
-import rehypeShiki from "@shikijs/rehype";
-import rehypeStringify from "rehype-stringify";
-import remarkParse from "remark-parse";
-import remarkRehype from "remark-rehype";
-import { unified } from "unified";
-
-import { transformerMetaHighlight } from "@shikijs/transformers";
-
 import { getFilesPaths } from "../utils/fs";
 
 const getNotes = async (): Promise<{ notes: Partial<Note>[] }> => {
@@ -85,39 +76,4 @@ const getNoteDirectory = async (
   };
 };
 
-// ! Should be in markdown.ts
-// ! rename to transformMarkdownFile
-// ! Type this and make it generic
-const transformMarkdownFile = async (pagePath: string) => {
-  const markdownWithMeta = fs.readFileSync(
-    `${decodeURIComponent(pagePath)}.md`,
-    "utf-8"
-  );
-
-  const { data: frontmatter, content } = matter(markdownWithMeta);
-
-  // console.log("content: ", content);
-
-  // ! This is needed for code syntax highlighting
-  // ! This allows for line highlighting but also requires a '.hightlighted' class to be in the index.css
-  const file = await unified()
-    .use(remarkParse) // Parse markdown
-    .use(remarkRehype, { allowDangerousHtml: true }) // Turn markdown into HTML, and allow raw HTML
-    .use(rehypeShiki, {
-      theme: "synthwave-84",
-      transformers: [transformerMetaHighlight()],
-    }) // Syntax highlighting
-    .use(rehypeStringify, { allowDangerousHtml: true }) // Turn HTML into string
-    .process(content);
-
-  const html = file.value;
-
-  console.log("html: ", html);
-
-  return {
-    frontmatter,
-    html,
-  };
-};
-
-export { getNotes, getNoteDirectory, transformMarkdownFile };
+export { getNotes, getNoteDirectory };
